# Optimizations and bug fixes for editor-devtools