**Changes**

 - The addon previously used an interesting method to hide the context menu when a button is clicked: hide the entire workspace and show it again. This is horribly slow on large projects because all the styles have to get recomputed. It now uses fake events to trick it into hiding (without access to the real Blockly and without using Blockly's builtin context menu handling, this seems to be the most consistent way to do this) which makes any context menu operation (copy, cut, paste, swap variable, show senders/receivers...) significantly faster.
   - "Copy block" before & after: [2.37 seconds](https://user-images.githubusercontent.com/33787854/103255960-fca7d180-4950-11eb-9e75-e6a5ce523d02.png) to [~200ms](https://user-images.githubusercontent.com/33787854/103256058-5e683b80-4951-11eb-9bfc-8073b8aff281.png) (Runtime sprite in https://scratch.mit.edu/projects/290745095/editor/)
   - "Show senders/receivers" before & after: [236ms](https://user-images.githubusercontent.com/33787854/103256452-fa467700-4952-11eb-85b4-258987c2c869.png) to [28ms](https://user-images.githubusercontent.com/33787854/103256490-12b69180-4953-11eb-8d83-a2fff433946e.png) (Player sprite of https://scratch.mit.edu/projects/60917032/editor/)
 - Swaps, cleanups, and cuts can be undone/redone with a single undo/redo instead of many.
 - Remove leftover `debugger;` and references to removed messages

**Tests**

it probably works fine